address feedback, move returnTip to compound
