+++ step-generation/src/__tests__/replaceTip.test.ts
@@ -19,7 +19,7 @@ import {
   DEFAULT_PIPETTE,
   PIPETTE_96,
 } from '../fixtures'
-import { replaceTip } from '../commandCreators/atomic/replaceTip'
+import { replaceTip } from '../commandCreators/compound/replaceTip'
 import { FIXED_TRASH_ID } from '../constants'
 import type { LabwareDefinition2 } from '@opentrons/shared-data'
 import type { InvariantContext, RobotState } from '../types'
