+++ step-generation/src/commandCreators/atomic/index.ts
@@ -21,10 +21,11 @@ import { moveLabware } from './moveLabware'
 import { moveToAddressableArea } from './moveToAddressableArea'
 import { moveToAddressableAreaForDropTip } from './moveToAddressableAreaForDropTip'
 import { moveToWell } from './moveToWell'
-import { replaceTip } from './replaceTip'
 import { setTemperature } from './setTemperature'
 import { touchTip } from './touchTip'
 import { waitForTemperature } from './waitForTemperature'
+import { pickUpTip } from './pickUpTip'
+
 export {
   absorbanceReaderCloseLid,
   absorbanceReaderInitialize,
@@ -34,9 +35,9 @@ export {
   aspirateInPlace,
   blowout,
   blowOutInPlace,
+  comment,
   configureForVolume,
   configureNozzleLayout,
-  comment,
   deactivateTemperature,
   delay,
   disengageMagnet,
@@ -49,7 +50,7 @@ export {
   moveToAddressableArea,
   moveToAddressableAreaForDropTip,
   moveToWell,
-  replaceTip,
+  pickUpTip,
   setTemperature,
   touchTip,
   waitForTemperature,
