+++ step-generation/src/commandCreators/compound/heaterShaker.ts
@@ -20,6 +20,7 @@ export const heaterShaker: CommandCreator<HeaterShakerArgs> = (
   invariantContext,
   prevRobotState
 ) => {
+
   if (args.moduleId == null) {
     return {
       errors: [errorCreators.missingModuleError()],
