+++ step-generation/src/commandCreators/compound/index.ts
@@ -2,8 +2,9 @@ export { absorbanceReaderCloseInitialize } from './absorbanceReaderCloseInitiali
 export { absorbanceReaderCloseRead } from './absorbanceReaderCloseRead'
 export { consolidate } from './consolidate'
 export { distribute } from './distribute'
+export { heaterShaker } from './heaterShaker'
 export { mix } from './mix'
+export { replaceTip } from './replaceTip'
 export { thermocyclerProfileStep } from './thermocyclerProfileStep'
 export { thermocyclerStateStep } from './thermocyclerStateStep'
 export { transfer } from './transfer'
-export { heaterShaker } from './heaterShaker'
