+++ step-generation/src/commandCreators/index.ts
@@ -1,13 +1,14 @@
 export {
-  transfer,
-  mix,
-  consolidate,
-  distribute,
-  thermocyclerProfileStep,
-  thermocyclerStateStep,
-  heaterShaker,
   absorbanceReaderCloseInitialize,
   absorbanceReaderCloseRead,
+  consolidate,
+  distribute,
+  heaterShaker,
+  mix,
+  replaceTip,
+  thermocyclerProfileStep,
+  thermocyclerStateStep,
+  transfer,
 } from './compound'
 
 export {
@@ -16,8 +17,8 @@ export {
   absorbanceReaderOpenLid,
   absorbanceReaderRead,
   aspirate,
-  waitForTemperature,
   blowout,
+  comment,
   deactivateTemperature,
   delay,
   disengageMagnet,
@@ -25,10 +26,9 @@ export {
   dropTip,
   dropTipInPlace,
   engageMagnet,
-  replaceTip,
-  setTemperature,
-  touchTip,
   moveLabware,
   moveToAddressableArea,
-  comment,
+  setTemperature,
+  touchTip,
+  waitForTemperature,
 } from './atomic'
