+++ step-generation/src/constants.ts
@@ -9,7 +9,11 @@ import {
   MAGNETIC_BLOCK_TYPE,
 } from '@opentrons/shared-data'
 
-import type { ModuleType, ModuleModel } from '@opentrons/shared-data'
+import type {
+  ModuleType,
+  ModuleModel,
+  AddressableOffsetVector,
+} from '@opentrons/shared-data'
 import type {
   MagneticModuleState,
   TemperatureModuleState,
@@ -73,10 +77,10 @@ export const MODULE_INITIAL_STATE_BY_TYPE: {
   [TEMPERATURE_MODULE_TYPE]: TEMPERATURE_MODULE_INITIAL_STATE,
   [THERMOCYCLER_MODULE_TYPE]: THERMOCYCLER_MODULE_INITIAL_STATE,
   [HEATERSHAKER_MODULE_TYPE]: HEATERSHAKER_MODULE_INITIAL_STATE,
-  //  TODO(jr, 6/24/24): add the initial state for absorabance reader
   [ABSORBANCE_READER_TYPE]: ABSORBANCE_READER_INITIAL_STATE,
   [MAGNETIC_BLOCK_TYPE]: MAGNETIC_BLOCK_INITIAL_STATE,
 }
 export const OT_2_TRASH_DEF_URI = 'opentrons/opentrons_1_trash_1100ml_fixed/1'
 export const FLEX_TRASH_DEF_URI = 'opentrons/opentrons_1_trash_3200ml_fixed/1'
 export const COLUMN_4_SLOTS = ['A4', 'B4', 'C4', 'D4']
+export const ZERO_OFFSET: AddressableOffsetVector = { x: 0, y: 0, z: 0 }
