+++ step-generation/src/utils/wasteChuteCommandsUtil.ts
@@ -5,6 +5,7 @@ import {
   dropTipInPlace,
   moveToAddressableArea,
 } from '../commandCreators/atomic'
+import { ZERO_OFFSET } from '../constants'
 import { curryCommandCreator } from './curryCommandCreator'
 import type { AddressableAreaName } from '@opentrons/shared-data'
 import type { RobotState, CurriedCommandCreator } from '../types'
@@ -35,7 +36,7 @@ export const wasteChuteCommandsUtil = (
     volume,
     flowRate,
   } = args
-  const offset = { x: 0, y: 0, z: 0 }
+  const offset = ZERO_OFFSET
   let commands: CurriedCommandCreator[] = []
   switch (type) {
     case 'dropTip': {
